Update bug_report.md
This new version of the bug report template adds more fields and is formatted in a way that is more systematic and matches up with an FMEA-style analysis of the issue. It also includes some guidance on communication with the team and and impact to production based on the risk flag of the issue.